move query utils out into separate file
